requirements.txt
amqp==1.4.6
anyjson==0.3.3
bcrypt==1.1.1
billiard==3.3.0.19
cffi==0.9.2
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.7.6
django-autocomplete-light==2.0.9
django-autoslug==1.7.2
django-cities-light==3.1.2
django-common-helpers==0.6.4
django-cron==0.3.6
django-jsonfield==0.9.13
django-post-office==1.1.1
django-registration-redux==1.1
django-tables2==0.15.0
django-toolbelt==0.0.1
gunicorn==19.3.0
jsonfield==1.0.3
kombu==3.0.24
psycopg2==2.4.5
pycparser==2.10
pytz==2015.2
six==1.9.0
South==1.0.2
static3==0.5.1
Unidecode==0.4.17
